extract beam-tcp-server for test
